FIX: Mudado tipo do ID de UCSAP para short

+++ src/projetoredes/unicast/UCSAP.java
@@ -2,4 +2,4 @@ package projetoredes.unicast;
 
 // A classe 'record' automaticamente gera construtores, métodos 'equals', 
 // 'hashCode' e 'toString' com base nos campos definidos.
-public record UCSAP(int id, String host, int port) {}
+public record UCSAP(short id, String host, int port) {}
